requirements.txt
uvicorn
fastapi
pydantic
psycopg2-binary
email-validator
